(chore) remove stripe meteor methods
