+++ imports/plugins/included/payments-stripe/server/index.js
@@ -1,5 +1,4 @@
 import "./i18n";
-import "./methods/stripe";
 import "./methods/stripe-connect";
 import "./startup/startup";
